GUI files structure refactor
Changes:
- split insert_services, manipulate_cities, manipulate_regions and update_services content to a separate files
- Add docstrings and silent other pylint errors
