+++ platform_management/gui/basics.py
@@ -1,4 +1,3 @@
-# pylint: disable=c-extension-no-member
 """
 Common Qt widgets to use in GUI application.
 """
@@ -16,6 +15,10 @@ logger = logger.bind(name="basics")
 def check_geometry_correctness(
     geometry_geojson: Optional[str], conn: "psycopg2.connection"
 ) -> Optional[Tuple[float, float, str]]:
+    """Check the correctness of the geometry by passing it to the PostGIS ST_GeomFromGeoJSON.
+
+    Return centroid (latitude, longitude, geometry type) if geometry is correct, None otherwise.
+    """
     if geometry_geojson is None:
         return None
     try:
@@ -36,6 +39,8 @@ def check_geometry_correctness(
 
 
 class ColorizingLine(QtWidgets.QLineEdit):
+    """Text line with an ability to set a hook for a change on focusOut event."""
+
     def __init__(
         self,
         callback: Callable[[Optional[QtWidgets.QLineEdit], Optional[str]], None],
@@ -61,6 +66,8 @@ class ColorizingLine(QtWidgets.QLineEdit):
 
 
 class ColorizingComboBox(QtWidgets.QComboBox):
+    """Combo box with an ability to set a hook for the change event"""
+
     def __init__(
         self,
         callback: Callable[[Optional[QtWidgets.QComboBox], Optional[int]], None],
@@ -83,6 +90,8 @@ class ColorizingComboBox(QtWidgets.QComboBox):
 
 
 class CheckableTableView(QtWidgets.QTableView):
+    """Table view with an ability to disable some of the rows from being inserted in the database."""
+
     colorTable = NamedTuple("ColorTable", [("on", QtGui.QColor), ("off", QtGui.QColor)])(
         QtGui.QColor(152, 224, 173), QtGui.QColor(248, 161, 164)
     )  # type: ignore
@@ -121,6 +130,7 @@ class CheckableTableView(QtWidgets.QTableView):
             super().keyPressEvent(event)
 
     def toggle_row(self, row: int) -> None:
+        """Change row uploading status."""
         item_index = self.model().index(row, 0)
         item = self.model().data(item_index)
         self.model().setData(item_index, "-" if item == "+" else "+")
@@ -131,20 +141,25 @@ class CheckableTableView(QtWidgets.QTableView):
         )
 
     def turn_row_on(self, row: int) -> None:
+        """Enable row to be uploaded."""
         item_index = self.model().index(row, 0)
         self.model().setData(item_index, "+")
         self.model().setData(item_index, CheckableTableView.colorTable.on, QtCore.Qt.BackgroundRole)
 
     def turn_row_off(self, row: int) -> None:
+        """Disable row from being uploaded."""
         item_index = self.model().index(row, 0)
         self.model().setData(item_index, "-")
         self.model().setData(item_index, CheckableTableView.colorTable.off, QtCore.Qt.BackgroundRole)
 
     def is_turned_on(self, row: int) -> bool:
+        """Return True if the row is not disabled."""
         return self.model().itemData(self.model().index(row, 0)) == "+"
 
 
 class DropPushButton(QtWidgets.QPushButton):
+    """Button with a drag-and-drop interface for files."""
+
     def __init__(
         self, text: str, formats: List[str], callback: Callable[[str], None], parent: Optional[QtWidgets.QWidget] = None
     ):
@@ -168,7 +183,9 @@ class DropPushButton(QtWidgets.QPushButton):
 
 
 class ColoringTableWidget(QtWidgets.QTableWidget):
-    def __init__(
+    """Table with an ability to set a hook on a cell change"""
+
+    def __init__(  # pylint: disable=too-many-arguments
         self,
         data: List[Sequence[Any]],
         labels: List[str],
@@ -215,13 +232,17 @@ class ColoringTableWidget(QtWidgets.QTableWidget):
         return super().dataChanged(top_left, bottom_right, roles=roles)
 
     def disable_triggers(self) -> None:
+        """Disable calling callback on cell change."""
         self._initialized = False
 
     def enable_triggers(self) -> None:
+        """Enable calling callback on cell change."""
         self._initialized = True
 
 
 class GeometryShow(QtWidgets.QDialog):
+    """Geometry display window with option to copy GeoJSON data."""
+
     def __init__(self, geometry: str, parent: Optional[QtWidgets.QWidget] = None):
         super().__init__(parent=parent)
         self.window().setWindowTitle("Просмотр геометрии")
