+++ platform_management/gui/insert_services/__init__.py
@@ -1,0 +1,3 @@
+"""Services insertion GUI logic is located here."""
+
+from .insert_services import ServicesInsertionWindow
