+++ platform_management/gui/insert_services/logging.py
@@ -1,0 +1,4 @@
+"""Logging for services insertion process is defined here."""
+from loguru import logger
+
+logger = logger.bind(name="gui_insert_services")
