+++ platform_management/gui/manipulate_cities/__init__.py
@@ -1,0 +1,2 @@
+"""Platform cities GUI manupulation functions are located here."""
+from .manipulate_cities import CitiesWindow
