+++ platform_management/gui/manipulate_cities/logging.py
@@ -1,0 +1,4 @@
+"""Logging for cities manipulation process is defined here."""
+from loguru import logger
+
+logger = logger.bind(name="gui_manipulate_cities")
