+++ platform_management/gui/manipulate_regions/__init__.py
@@ -1,0 +1,2 @@
+"""Regions manupulation GUI functions are located here."""
+from .manipulate_regions import RegionsWindow
