+++ platform_management/gui/manipulate_regions/logging.py
@@ -1,0 +1,4 @@
+"""Logging for regions manipulation process is defined here."""
+from loguru import logger
+
+logger = logger.bind(name="gui_manipulate_regions")
