+++ platform_management/gui/update_services/__init__.py
@@ -1,0 +1,6 @@
+"""Service updation GUI logic is defined here."""
+from .building_creation import BuildingCreationWidget
+from .geometry_show import GeometryShowWidget
+from .physical_object_creation import PhysicalObjectCreationWidget
+from .platform_services_table import PlatformServicesTableWidget
+from .update_services import ServicesUpdatingWindow
