+++ platform_management/gui/update_services/logging.py
@@ -1,0 +1,4 @@
+"""Logging for services updation process is defined here."""
+from loguru import logger
+
+logger = logger.bind(name="gui_update_services")
